server/pyproject.toml
[project]
name = "eudai-server"
version = "0.1.0"
description = "Server for Eudai, a conversational AI assistant designed to help users offload tasks and clarify their thoughts in order to achieve 'eudaimonia', or optimal human flourishing — without sacrificing their data, privacy, or autonomy."
authors = ["Bethany Fannin <contact@bethanyfannin.com"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
keywords = [
    "eudai",
    "eudaimonia",
    "conversational-ai",
    "task-assistant",
    "privacy-focused",
    "data-privacy",
    "autonomy",
    "productivity",
    "personal-knowledge-management",
    "knowledge-management",
    "self-improvement",
    "mental-wellbeing",
    "pkm",
    "ai",
    "llm",
    "ollama",
    "deepseek",
    "fastapi"
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Technology Enthusiasts",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.13",
    "Largely Maintained :: Yes",
    "Operating System :: OS Independent",
    "Topic :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
    "Framework :: FastAPI",
    "Framework :: Poetry"
]
dependencies = [
    "fastapi>=0.100.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.3",
    "uvicorn[standard]>=0.28.0",
    "ollama>=0.1.0",
    "deepseek>=0.1.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.6.0",
    "rich>=13.5.2",
    "colorama>=0.4.6"
]
optional-dependencies.dev = [
    "Pygments>=2.10.0",
    "black>=21.10b0",
    "coverage[toml]>=6.2",
    "darglint>=1.8.1",
    "flake8>=4.0.1",
    "flake8-bandit>=2.1.2",
    "flake8-bugbear>=21.9.2",
    "flake8-docstrings>=1.6.0",
    "flake8-rst-docstrings>=0.2.5",
    "furo>=2021.11.12",
    "isort>=5.10.1",
    "mypy>=0.930",
    "pep8-naming>=0.12.1",
    "pre-commit>=2.16.0",
    "pre-commit-hooks>=4.1.0",
    "pytest>=6.2.5",
    "pyupgrade>=2.29.1",
    "safety>=1.10.3",
    "sphinx>=4.3.2",
    "sphinx-autobuild>=2021.3.14",
    "sphinx-click>=3.0.2",
    "typeguard>=2.13.3",
    "xdoctest[colors]>=0.15.10",
    "myst-parser>=0.16.1"
    ]
urls = { "Changelog" = "https://github.com/bjf5201/eudai/releases", "Homepage" = "https://github.com/bjf5201/eudai/blob/main/server/README.md", "Repository" = "https://github.com/bjf5201/eudai/blob/main/server", "Documentation" = "https://eudai.readthedocs.io" }
scripts = { "eudai-server"="eudai-server.__main__:main"}

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]


[tool.coverage.run]
branch = true
source = ["eudai-server", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
